requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
idna==2.9
kiwisolver==1.1.0
matplotlib==3.2.0
numpy==1.18.1
pandas==1.0.1
plotly==4.5.4
prompt-toolkit==1.0.14
pyfiglet==0.8.post1
Pygments==2.6.1
PyInquirer==1.0.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.12.1
